backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "event-platform-backend"
version = "0.1.0"
description = "FastAPI backend for the event platform"
requires-python = ">=3.11"
dependencies = [
  "alembic>=1.13,<2.0",
  "asyncpg>=0.29,<1.0",
  "fastapi>=0.115,<1.0",
  "httpx>=0.27,<1.0",
  "passlib[bcrypt]>=1.7,<2.0",
  "pydantic-settings>=2.4,<3.0",
  "python-jose[cryptography]>=3.3,<4.0",
  "python-multipart>=0.0.9,<1.0",
  "sqlalchemy[asyncio]>=2.0,<3.0",
  "uvicorn[standard]>=0.30,<1.0"
]

[project.optional-dependencies]
dev = [
  "mypy>=1.11,<2.0",
  "pytest>=8.0,<9.0",
  "pytest-asyncio>=0.23,<1.0",
  "ruff>=0.6,<1.0"
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
